feat: finish purchase items service
Fix return from get all purchased items;
Insert routes into swagger file;

+++ src/purchase-items/purchase-items.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { PurchaseItems } from './entities/purchase-items.entity';
 import { InjectKnex, Knex } from 'nestjs-knex';
-import { PRODUCT_TABLE, PURCHASE_ITEMS_TABLE } from '../constants';
+import { PURCHASE_ITEMS_TABLE } from '../constants';
 import { PurchaseItemsInput } from './dto/purchase-items-input.dto';
 
 @Injectable()
@@ -13,8 +13,7 @@ export class PurchaseItemsRepository {
   
   async selectPurchaseItems(): Promise<PurchaseItems[] | null> {
     return await this.knex<PurchaseItems>(PURCHASE_ITEMS_TABLE)
-      .select<PurchaseItems[]>()
-      .leftJoin(PRODUCT_TABLE, `${PRODUCT_TABLE}.id`, `${PURCHASE_ITEMS_TABLE}.product`);
+      .select<PurchaseItems[]>();
   }
 
   async insertPurchaseItems(purchaseItemsValues: PurchaseItemsInput): Promise<PurchaseItems> {
@@ -31,10 +30,10 @@ export class PurchaseItemsRepository {
     await this.knex<PurchaseItems>(PURCHASE_ITEMS_TABLE).delete().where({ id });
   }
 
-  async deletePurchaseItemsByDate(id: string, filterDate: string) {
+  async deletePurchaseItemsByDate(userId: string, filterDate: string) {
     await this.knex<PurchaseItems>(PURCHASE_ITEMS_TABLE)
       .delete()
-      .where({ id })
+      .where({ user: userId })
       .andWhere('purchaseDate', '<', filterDate);
   }
 }
